requirements.txt
streamlit==1.24.0
openai-whisper==20230314
torch
tqdm
ffmpeg-python
Werkzeug==2.0.1
pydub==0.25.1
